pyproject.toml
[tool.poetry]
name = "fs.googledrivefs"
packages = [
    { include = "fs"}
]
version = "2.0.0"
description = "Pyfilesystem2 implementation for Google Drive"
authors = ["Rehan Khwaja <rehan@khwaja.name>"]
license = "MIT"
homepage = "https://github.com/rkhwaja/fs.googledrivefs"
keywords = ["filesystem", "Pyfilesystem2", "google"]
readme = "README.md"
classifiers = [
	"License :: OSI Approved :: MIT License",
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: Developers",
	"Programming Language :: Python :: 3.6",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9"
]

[tool.poetry.dependencies]
python = ">=3.6"
fs = ">=2.4.10"
google-api-python-client = ">=1.6.3"
google-auth = ">=1.5.1"

[tool.poetry.dev-dependencies]
pytest = ">=3.10"
pytest-timeout = ">=1.3.3"
requests-oauthlib = "^1.0"
pyperclip = "^1.7"
pylint = ">=2.5.3"
pylint-quotes = "^0.2.1"
pytest-cov = "^2.10.1"
pyngrok = "^1.4"
pytest-localserver = "^0.5.0"

[tool.poetry.plugins] # Optional super table

[tool.poetry.plugins."fs.opener"]
"googledrive" = "fs.googledrivefs.opener:GoogleDriveFSOpener"

[tool.pylint.master]
load-plugins = "pylint_quotes"

[tool.pylint.format]
indent-string = "\t"
max-line-length = "200"
string-quote = "single-avoid-escape"

[tool.pylint.messages_control]
# no-member to remove false positives with objects constructed via google's discovery protocol
disable = [
	"bad-continuation",
	"fixme",
	"invalid-name",
	"len-as-condition",
	"line-too-long",
	"logging-format-interpolation",
	"logging-fstring-interpolation",
	"missing-docstring",
	"no-member"
]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
